Move hooks to proper directories
Also fix some imports not to point directly to the files
