+++ ui/src/hooks/index.ts
@@ -3,18 +3,12 @@ export * from './lines';
 export * from './mutationHook';
 export * from './redux';
 export * from './routes';
-export * from './search/useSearch';
-export * from './search/useSearchQueryParser';
-export * from './search/useSearchResults';
+export * from './search';
 export * from './stops';
 export * from './ui';
 export * from './urlQuery';
 export * from './useAsyncQuery';
 export * from './useCheckValidityAndPriorityConflicts';
-export * from './useChooseLineDropdown';
-export * from './useChooseRouteDropdown';
-export * from './useExtractRouteFromFeature';
-export * from './useFilterStops';
 export * from './usePagination';
 export * from './useShowRoutesOnModal';
 export * from './via';
