+++ ui/src/hooks/routes/index.ts
@@ -1,6 +1,8 @@
+export * from './useCreateRoute';
 export * from './useDeleteRoute';
 export * from './useEditRouteGeometry';
 export * from './useEditRouteJourneyPattern';
 export * from './useEditRouteMetadata';
+export * from './useExtractRouteFromFeature';
 export * from './useGetDisplayedRoutes';
 export * from './useValidateRoute';
