+++ ui/src/hooks/stops/index.ts
@@ -1,6 +1,7 @@
 export * from './useCreateStop';
 export * from './useDeleteStop';
 export * from './useEditStop';
+export * from './useFilterStops';
 export * from './useGetStopLinkAndDirection';
 export * from './useMapStops';
 export * from './utils';
